moved data.rs in favour of data/mod.rs and extracted mod witness_trees
